Add configuration object.
This changes the implementation to use a very basic conf file for
configuration. I think it's a little more flexible for installed vs
portable usage. As part of this change many paths are no longer hard
coded and instead are based on the root and meta-root configuration
values.

The config file is a pretty standard format. Lines beginning with # are
comments. Blank lines are ignored. Other lines are key=value pairs.

I've also moved the default meta-root to .meta (was .data).